chore(ui): refactor influxdb utility functions
