+++ ui/src/admin/containers/influxdb/RolesPage.tsx
@@ -21,8 +21,8 @@ import FancyScrollbar from 'src/shared/components/FancyScrollbar'
 import EmptyRow from 'src/admin/components/EmptyRow'
 import RoleRow from 'src/admin/components/RoleRow'
 import {useCallback} from 'react'
-import allOrParticularSelection from './util/allOrParticularSelection'
-import {computeEntitiesDBPermissions} from './util/computeEffectiveDBPermissions'
+import allOrParticularSelection from '../../util/allOrParticularSelection'
+import {computeEntitiesDBPermissions} from '../../util/computeEffectiveDBPermissions'
 import useDebounce from 'src/utils/useDebounce'
 import useChangeEffect from 'src/utils/useChangeEffect'
 import {ComponentSize, MultiSelectDropdown, SlideToggle} from 'src/reusable_ui'
