+++ ui/src/admin/util/allOrParticularSelection.ts
